Add a .gitattributes file.
This ensures all text files are properly marked as such from git's
perspective. This will allow it to determine whether it should convert
line endings to the platform's native endings on checkout, for a given
file.

The Python-specific parts were sourced from here
https://github.com/alexkaratarakis/gitattributes/blob/master/Python.gitattributes
For all other files, it lets git's heuristics determine whether it is
a text file.