Update Dockerfile base image and installation commands
The base image in the Dockerfile has been updated to python:3.11-slim-bookworm from alpine:3 for a more tailored environment. The initial installation commands using apk have been replaced with apt commands. The apk commands were removed as they are not compatible with the new base image. This fixes compt issues with ujson